fix(lang_class): cast ACC_PUBLIC (i32) to u16 for MethodMetadata.access_flags in LayoutProcessor overload-resolution test
Pre-existing type mismatch (E0308) from a same-day dev commit adding
lang_class.rs's example/LayoutProcessor overload-resolution test, unrelated
to this branch's WildFly work -- was blocking 'cargo test' compilation for
the whole native-builtins crate. ACC_PUBLIC is imported as the i32-typed
ACC_PUBLIC_I32 alias; MethodMetadata.access_flags is u16. Fixed with an
explicit as-cast at the two test-only construction sites (0x0001 fits u16
trivially, no behavior change).

+++ native-builtins/src/lang_class.rs
@@ -19487,14 +19487,14 @@ Implementation-Title: opensaml-core-api\r\n\
                 MethodMetadata {
                     name: "<init>".to_string(),
                     descriptor: "()V".to_string(),
-                    access_flags: ACC_PUBLIC,
+                    access_flags: ACC_PUBLIC as u16,
                     declaring_class_id: cid,
                     exceptions: Vec::new(),
                 },
                 MethodMetadata {
                     name: "<init>".to_string(),
                     descriptor: "(Ljava/lang/String;II)V".to_string(),
-                    access_flags: ACC_PUBLIC,
+                    access_flags: ACC_PUBLIC as u16,
                     declaring_class_id: cid,
                     exceptions: Vec::new(),
                 },
